Add a Stack parameter to REST API Fixes #273
