+++ go/models/code_generator.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path/filepath"
 	"sort"
 	"strings"
 )
@@ -255,94 +254,5 @@ func MultiCodeGenerator(
 
 		defer file.Close()
 		fmt.Fprint(file, code)
-	}
-}
-
-// RemoveGeneratedGongFiles generates the setup file for the gorm
-func RemoveGeneratedGongFiles(
-	RelativePkgPath string) {
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
-	}
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong_coder.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
-	}
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong_ast.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
-	}
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong_serialize.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
-	}
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong_marshall.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
-	}
-
-	{
-		// relative to the models package, swith to ./controlers package
-		filename := filepath.Join(RelativePkgPath, "gong_graph.go")
-
-		// we should use go generate
-		log.Println("removing file : " + filename)
-
-		if err := os.Remove(filename); err != nil {
-			if os.IsExist(err) {
-				log.Fatalf("Unable to remove %s", filename)
-			}
-		}
 	}
 }
