+++ test/go/controllers/AstructBstruct2Use_CRUDs.go
@@ -52,6 +52,19 @@ func GetAstructBstruct2Uses(c *gin.Context) {
 
 	// source slice
 	var astructbstruct2useDBs []orm.AstructBstruct2UseDB
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET all params", stackParam)
+		}
+	}
+
 	query := db.Find(&astructbstruct2useDBs)
 	if query.Error != nil {
 		var returnError GenericError
@@ -96,7 +109,6 @@ func GetAstructBstruct2Uses(c *gin.Context) {
 //	Responses:
 //	  200: nodeDBResponse
 func PostAstructBstruct2Use(c *gin.Context) {
-	db := orm.BackRepo.BackRepoAstructBstruct2Use.GetDB()
 
 	// Validate input
 	var input orm.AstructBstruct2UseAPI
@@ -116,6 +128,7 @@ func PostAstructBstruct2Use(c *gin.Context) {
 	astructbstruct2useDB.AstructBstruct2UsePointersEnconding = input.AstructBstruct2UsePointersEnconding
 	astructbstruct2useDB.CopyBasicFieldsFromAstructBstruct2Use(&input.AstructBstruct2Use)
 
+	db := orm.BackRepo.BackRepoAstructBstruct2Use.GetDB()
 	query := db.Create(&astructbstruct2useDB)
 	if query.Error != nil {
 		var returnError GenericError
@@ -152,6 +165,19 @@ func PostAstructBstruct2Use(c *gin.Context) {
 //
 //	200: astructbstruct2useDBResponse
 func GetAstructBstruct2Use(c *gin.Context) {
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET params", stackParam)
+		}
+	}
+
 	db := orm.BackRepo.BackRepoAstructBstruct2Use.GetDB()
 
 	// Get astructbstruct2useDB in DB
@@ -184,6 +210,15 @@ func GetAstructBstruct2Use(c *gin.Context) {
 //
 //	200: astructbstruct2useDBResponse
 func UpdateAstructBstruct2Use(c *gin.Context) {
+
+	// Validate input
+	var input orm.AstructBstruct2UseAPI
+	if err := c.ShouldBindJSON(&input); err != nil {
+		log.Println(err.Error())
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	
 	db := orm.BackRepo.BackRepoAstructBstruct2Use.GetDB()
 
 	// Get model if exist
@@ -198,14 +233,6 @@ func UpdateAstructBstruct2Use(c *gin.Context) {
 		returnError.Body.Message = query.Error.Error()
 		log.Println(query.Error.Error())
 		c.JSON(http.StatusBadRequest, returnError.Body)
-		return
-	}
-
-	// Validate input
-	var input orm.AstructBstruct2UseAPI
-	if err := c.ShouldBindJSON(&input); err != nil {
-		log.Println(err.Error())
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
