+++ test/go/controllers/AstructBstructUse_CRUDs.go
@@ -52,6 +52,19 @@ func GetAstructBstructUses(c *gin.Context) {
 
 	// source slice
 	var astructbstructuseDBs []orm.AstructBstructUseDB
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET all params", stackParam)
+		}
+	}
+
 	query := db.Find(&astructbstructuseDBs)
 	if query.Error != nil {
 		var returnError GenericError
@@ -96,7 +109,6 @@ func GetAstructBstructUses(c *gin.Context) {
 //	Responses:
 //	  200: nodeDBResponse
 func PostAstructBstructUse(c *gin.Context) {
-	db := orm.BackRepo.BackRepoAstructBstructUse.GetDB()
 
 	// Validate input
 	var input orm.AstructBstructUseAPI
@@ -116,6 +128,7 @@ func PostAstructBstructUse(c *gin.Context) {
 	astructbstructuseDB.AstructBstructUsePointersEnconding = input.AstructBstructUsePointersEnconding
 	astructbstructuseDB.CopyBasicFieldsFromAstructBstructUse(&input.AstructBstructUse)
 
+	db := orm.BackRepo.BackRepoAstructBstructUse.GetDB()
 	query := db.Create(&astructbstructuseDB)
 	if query.Error != nil {
 		var returnError GenericError
@@ -152,6 +165,19 @@ func PostAstructBstructUse(c *gin.Context) {
 //
 //	200: astructbstructuseDBResponse
 func GetAstructBstructUse(c *gin.Context) {
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET params", stackParam)
+		}
+	}
+
 	db := orm.BackRepo.BackRepoAstructBstructUse.GetDB()
 
 	// Get astructbstructuseDB in DB
@@ -184,6 +210,15 @@ func GetAstructBstructUse(c *gin.Context) {
 //
 //	200: astructbstructuseDBResponse
 func UpdateAstructBstructUse(c *gin.Context) {
+
+	// Validate input
+	var input orm.AstructBstructUseAPI
+	if err := c.ShouldBindJSON(&input); err != nil {
+		log.Println(err.Error())
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	
 	db := orm.BackRepo.BackRepoAstructBstructUse.GetDB()
 
 	// Get model if exist
@@ -198,14 +233,6 @@ func UpdateAstructBstructUse(c *gin.Context) {
 		returnError.Body.Message = query.Error.Error()
 		log.Println(query.Error.Error())
 		c.JSON(http.StatusBadRequest, returnError.Body)
-		return
-	}
-
-	// Validate input
-	var input orm.AstructBstructUseAPI
-	if err := c.ShouldBindJSON(&input); err != nil {
-		log.Println(err.Error())
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
