+++ test/go/controllers/Bstruct_CRUDs.go
@@ -52,6 +52,19 @@ func GetBstructs(c *gin.Context) {
 
 	// source slice
 	var bstructDBs []orm.BstructDB
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET all params", stackParam)
+		}
+	}
+
 	query := db.Find(&bstructDBs)
 	if query.Error != nil {
 		var returnError GenericError
@@ -96,7 +109,6 @@ func GetBstructs(c *gin.Context) {
 //	Responses:
 //	  200: nodeDBResponse
 func PostBstruct(c *gin.Context) {
-	db := orm.BackRepo.BackRepoBstruct.GetDB()
 
 	// Validate input
 	var input orm.BstructAPI
@@ -116,6 +128,7 @@ func PostBstruct(c *gin.Context) {
 	bstructDB.BstructPointersEnconding = input.BstructPointersEnconding
 	bstructDB.CopyBasicFieldsFromBstruct(&input.Bstruct)
 
+	db := orm.BackRepo.BackRepoBstruct.GetDB()
 	query := db.Create(&bstructDB)
 	if query.Error != nil {
 		var returnError GenericError
@@ -152,6 +165,19 @@ func PostBstruct(c *gin.Context) {
 //
 //	200: bstructDBResponse
 func GetBstruct(c *gin.Context) {
+
+	// type Values map[string][]string
+	values := c.Request.URL.Query()
+	if len(values) == 1 {
+		value := values["stack"]
+		if len(value) == 1 {
+			// we have a single parameter
+			// we assume it is the stack
+			stackParam := value[0]
+			log.Println("GET params", stackParam)
+		}
+	}
+
 	db := orm.BackRepo.BackRepoBstruct.GetDB()
 
 	// Get bstructDB in DB
@@ -184,6 +210,15 @@ func GetBstruct(c *gin.Context) {
 //
 //	200: bstructDBResponse
 func UpdateBstruct(c *gin.Context) {
+
+	// Validate input
+	var input orm.BstructAPI
+	if err := c.ShouldBindJSON(&input); err != nil {
+		log.Println(err.Error())
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	
 	db := orm.BackRepo.BackRepoBstruct.GetDB()
 
 	// Get model if exist
@@ -198,14 +233,6 @@ func UpdateBstruct(c *gin.Context) {
 		returnError.Body.Message = query.Error.Error()
 		log.Println(query.Error.Error())
 		c.JSON(http.StatusBadRequest, returnError.Body)
-		return
-	}
-
-	// Validate input
-	var input orm.BstructAPI
-	if err := c.ShouldBindJSON(&input); err != nil {
-		log.Println(err.Error())
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
