+++ test/ng/projects/test/src/lib/astruct.service.ts
@@ -1,6 +1,6 @@
 // generated by ng_file_service_ts
 import { Injectable, Component, Inject } from '@angular/core';
-import { HttpClientModule, HttpParams } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { DOCUMENT, Location } from '@angular/common'
 
@@ -23,8 +23,7 @@ import { DstructDB } from './dstruct-db'
 export class AstructService {
 
   httpOptions = {
-    headers: new HttpHeaders({ 'Content-Type': 'application/json' }),
-    params: new HttpParams()
+    headers: new HttpHeaders({ 'Content-Type': 'application/json' })
   };
 
   // Kamar Raïmo: Adding a way to communicate between components that share information
@@ -32,7 +31,6 @@ export class AstructService {
   AstructServiceChanged: BehaviorSubject<string> = new BehaviorSubject("");
 
   private astructsUrl: string
-  params = new HttpParams();
 
   constructor(
     private http: HttpClient,
@@ -52,10 +50,7 @@ export class AstructService {
 
   /** GET astructs from the server */
   getAstructs(): Observable<AstructDB[]> {
-
-    this.params = this.params.set('stack', "A");
-
-    return this.http.get<AstructDB[]>(this.astructsUrl, { params: this.params })
+    return this.http.get<AstructDB[]>(this.astructsUrl)
       .pipe(
         tap(_ => this.log('fetched astructs')),
         catchError(this.handleError<AstructDB[]>('getAstructs', []))
@@ -64,10 +59,8 @@ export class AstructService {
 
   /** GET astruct by id. Will 404 if id not found */
   getAstruct(id: number): Observable<AstructDB> {
-    this.params = this.params.set('stack', "A")
-
     const url = `${this.astructsUrl}/${id}`;
-    return this.http.get<AstructDB>(url, { params: this.params }).pipe(
+    return this.http.get<AstructDB>(url).pipe(
       tap(_ => this.log(`fetched astruct id=${id}`)),
       catchError(this.handleError<AstructDB>(`getAstruct id=${id}`))
     );
@@ -77,7 +70,6 @@ export class AstructService {
 
   /** POST: add a new astruct to the server */
   postAstruct(astructdb: AstructDB): Observable<AstructDB> {
-    astructdb.Stack = "A"
 
     // insertion point for reset of pointers and reverse pointers (to avoid circular JSON)
     astructdb.Bstruct = new BstructDB
@@ -97,11 +89,7 @@ export class AstructService {
     let _Astruct_Anarrayofa_reverse = astructdb.Astruct_Anarrayofa_reverse
     astructdb.Astruct_Anarrayofa_reverse = new AstructDB
 
-    return this.http.post<AstructDB>(
-      this.astructsUrl,
-      astructdb,
-      this.httpOptions,
-    ).pipe(
+    return this.http.post<AstructDB>(this.astructsUrl, astructdb, this.httpOptions).pipe(
       tap(_ => {
         // insertion point for restoration of reverse pointers
         astructdb.Astruct_Anarrayofa_reverse = _Astruct_Anarrayofa_reverse
